refactor: move InfluxDB store and ReadingsService into internal/measurement
Moves the InfluxDB Client implementation into internal/measurement/store_influx.go
and ReadingsService into internal/measurement/service.go. A DeviceFinder interface
decouples measurement from device, satisfied by a thin bridge in sensors.
sensors/influx.go and sensors/readings_service.go become thin wrappers.

Closes #99
