+++ internal/sensors/model.go
@@ -25,7 +25,7 @@ var ErrDeviceNotFound          = device.ErrNotFound
 var ErrCodeNotFound            = errors.New("provisioning code not found")
 var ErrCodeAlreadyUsed         = errors.New("provisioning code already used")
 var ErrInvalidCommand          = peripheral.ErrInvalidCommand
-var ErrInfluxWrite             = errors.New("failed to persist reading")
+var ErrInfluxWrite             = measurement.ErrInfluxWrite
 var ErrPeripheralNotFound      = peripheral.ErrNotFound
 var ErrPeripheralAlreadyExists = peripheral.ErrAlreadyExists
 var ErrPeripheralPinInUse      = peripheral.ErrPinInUse
