+++ internal/sensors/readings_service.go
@@ -2,73 +2,37 @@ package sensors
 
 import (
 	"context"
-	"errors"
-	"fmt"
 	"log/slog"
-	"time"
 
-	"github.com/fishhub-oss/fishhub-server/internal/senml"
+	"github.com/fishhub-oss/fishhub-server/internal/device"
+	"github.com/fishhub-oss/fishhub-server/internal/measurement"
 )
 
-// ReadingsService orchestrates sensor reading operations.
-type ReadingsService struct {
-	devices DeviceStore
-	querier ReadingQuerier
-	writer  ReadingWriter
-	logger  *slog.Logger
+// ReadingsService is an alias for measurement.ReadingsService kept for backward compatibility.
+type ReadingsService = measurement.ReadingsService
+
+// deviceStoreFinderBridge bridges device.Store (returning (Device, error)) to
+// measurement.DeviceFinder (returning only error).
+type deviceStoreFinderBridge struct {
+	store DeviceStore
+}
+
+func (b *deviceStoreFinderBridge) FindByIDAndUserID(ctx context.Context, deviceID, userID string) error {
+	_, err := b.store.FindByIDAndUserID(ctx, deviceID, userID)
+	if err != nil {
+		return device.ErrNotFound
+	}
+	return nil
 }
 
 func NewReadingsService(devices DeviceStore, querier ReadingQuerier, writer ReadingWriter, logger *slog.Logger) *ReadingsService {
 	if logger == nil {
 		logger = slog.Default()
 	}
-	return &ReadingsService{devices: devices, querier: querier, writer: writer, logger: logger}
-}
-
-// Query verifies device ownership then fetches readings from InfluxDB.
-// Returns ErrDeviceNotFound unwrapped if the device does not exist or is not
-// owned by userID.
-func (s *ReadingsService) Query(ctx context.Context, userID string, q ReadingQuery) ([]ReadingPoint, error) {
-	if _, err := s.devices.FindByIDAndUserID(ctx, q.DeviceID, userID); err != nil {
-		if !errors.Is(err, ErrDeviceNotFound) {
-			s.logger.Error("query readings: find device", "device_id", q.DeviceID, "error", err)
-		}
-		return nil, err
-	}
-	points, err := s.querier.QueryReadings(ctx, q)
-	if err != nil {
-		s.logger.Error("query readings", "device_id", q.DeviceID, "error", err)
-		return nil, fmt.Errorf("query readings: %w", err)
-	}
-	return points, nil
-}
-
-// Write parses a SenML payload and writes the reading to InfluxDB.
-// If writer is nil the call is a no-op (InfluxDB not configured).
-func (s *ReadingsService) Write(ctx context.Context, deviceID, userID string, body []byte) error {
-	reading, err := senml.Parse(body)
-	if err != nil {
-		return err
-	}
-
-	s.logger.Info("reading received", "device_id", deviceID, "bytes", len(body))
-
-	if s.writer == nil {
-		return nil
-	}
-
-	fields := make(map[string]any, len(reading.Measurements))
-	for _, m := range reading.Measurements {
-		fields[m.Name] = m.Value
-	}
-	if err := s.writer.WriteReading(ctx, Reading{
-		DeviceID:     deviceID,
-		UserID:       userID,
-		Timestamp:    time.Unix(reading.BaseTime, 0).UTC(),
-		Measurements: fields,
-	}); err != nil {
-		s.logger.Error("influx write", "device_id", deviceID, "error", err)
-		return fmt.Errorf("%w: %w", ErrInfluxWrite, err)
-	}
-	return nil
+	return measurement.NewReadingsService(
+		&deviceStoreFinderBridge{store: devices},
+		querier,
+		writer,
+		logger,
+	)
 }
